Fix warning due to comparison of signed and unsigned values

+++ src/osdp_phy.c
@@ -297,7 +297,7 @@ int osdp_phy_check_packet(struct osdp_pd *pd, uint8_t *buf, int len,
 	}
 
 	if (pkt_len > OSDP_PACKET_BUF_SIZE ||
-	    pkt_len < sizeof(struct osdp_packet_header)) {
+	    (unsigned long)pkt_len < sizeof(struct osdp_packet_header)) {
 		pd->reply_id = REPLY_NAK;
 		pd->ephemeral_data[0] = OSDP_PD_NAK_CMD_LEN;
 		return OSDP_ERR_PKT_NACK;
